requirements.txt
aiofiles==24.1.0
aiogram==3.20.0.post0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
asgiref==3.8.1
asyncio==3.4.3
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
distlib==0.3.9
Django==5.1.6
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
dnspython==2.7.0
filelock==3.17.0
frozenlist==1.6.0
idna==3.10
magic-filter==1.0.12
motor==3.7.1
multidict==6.4.3
pillow==11.1.0
pip==25.0.1
platformdirs==4.3.6
propcache==0.3.1
psycopg2==2.9.10
psycopg2-binary==2.9.10
pydantic==2.11.4
pydantic_core==2.33.2
PyJWT==2.10.1
pymongo==4.13.0
python-dotenv==1.1.0
reportlab==4.4.3
sqlparse==0.5.3
typing_extensions==4.13.2
typing-inspection==0.4.0
tzdata==2025.1
virtualenv==20.29.2
yarl==1.20.0
